Rename package from "gpn" to "toron".
Most of my packages have had very plain, utilitarian names but
I thought it would be nice to use a more metaphorical name for
this project.

A key feature of this tool kit is to reconcile non-equi-joinable
datasets that have arbitrary levels of granularity. A topological
soliton (or toron) serves an analogous function:

    A topological soliton or "toron" occurs when two adjoining
    structures or spaces are in some way "out of phase" with each
    other in ways that make a seamless transition between them
    impossible.

    <https://en.wikipedia.org/wiki/Topological_soliton>