Change column types from PYTEXT to TEXT_JSON.
This is a move away from the custom serialization code and back
towards the more standard JSON serialization. While JSON is more
limited, it's probably best to stick with an established, more
familiar solution. Doing this allows Toron to take advantage
of the increasingly common JSON functions for column CHECK
constraints.