Add support for schema_version handling.
In the future, it may be necessary to alter the schema of node
database files. This change provides a mechanism for checking
the schema version to provide useful feedback to the user and
prevent any data loss or corruption that could occur if Toron
were to manipulate an unsupported format.