Remove `cache=shared` option from in-memory databases.
Before allowing this type of access, the connect() function needs
to check if the database exists. But currently, the function only
checks for files on disk using os.path.exists().

More planning is needed to determine if shared-cache Nodes are even
a good idea. If compelling use cases arise in the future, they can
be added back.