Set foreign key references for `element_id` to DEFERRED.
This relaxes the necessary order of operations when working on the
`element`, `element_weight`, and `relation` tables when can happen
when the dataset must be coarsened.

NOTE: When coarsening the dataset, the `element_weight` and `relation`
records need to be handled with care so these references should avoid
using an ON DELETE CASCADE clause to mitigate accidental data loss.