Update get_raw_quantities() to use `userfunc` context manager. And remove accepts_json_input() and related tests.
The `userfunc` context manager creates AND REMOVES user-
defined functions. It is no longer necessary to have the
`accepts_json_input` weapper class.

Documentation for the sqlite3 Connection.create_function()
method now say that the *func* parameter may be "Set to None
to remove an existing SQL function." BUT PRIOR TO PYTHON 3.10.5,
SQL FUNCTION REMOVAL WAS UNDOCUMENTED! The new language was
added July 26, 2022 (python/cpython@5012bed) and is related
to Issue 95235 (python/cpython#95235).