Test user-defined SQL func and unexhausted generator.
The removal of user-defined SQL functions was added in commit 014568ae.
But as it turns out, these functions cannot be removed arbitrarily.
For example, if you make a query that uses a user-defined function,
you cannot remove the function until the query results have been fully
consumed. This creates problems in Toron when query results are only
partially consumed--raises an sqlite3.OperationalError exception.