Add BitList to use for "mapping_level" column values.
The "mapping_level" column stores bitmasks that encode the
level of precision at which a particular relation was made.

Currently the "mapping_level" column uses a SQLite INTEGER
type. Since the INTEGER type cannot hold a bitmask greater
than 64 bits, it cannot be used to encode mapping levels that
contain more than 64 index/element columns.

The new BitList class can support bitmasks of arbitrary length
which can be stored as BLOB types in SQLite. This will enable
"mapping_level" values to encode for more than 64 columns. This
will be added to the SQL schema in the future.