Modify 'relation' table--add 'relation_value' and update 'proportion'.
To keep relations from breaking when a dataset is refined or coarsened,
it's necessary to store the original weight value since the proportion
alone is inadequate.

When adding a relation, the weight values will be included as-is and
the proportion will be calculated afterward. With this change in how
proportion is used, the NOT NULL constraint is removed. The proportions
must be calculated *after* the mapping is fully loaded and allowing
NULL values on the initial INSERT seem more straightforward than
inserting a temporary dummy value.