Add support for deleting ambiguous relations when removing a column.
When deleting an index column, the default behavior should
be to raise an error and cancel the column deleting. But if
a user indicates that edges do not have to be preserved in
their current state (by passing `preserve_edges=False`), then
column removal can continue and unrepresentable relations
will be deleted.