Update BitFlags.from_bytes() comment to avoid confusion.
Talking about endianness could be misleading. BitFlags are left-aligned
(like little-endian numbers) but they are a sequence of flags which do
not actually represent a number. In contrast, the binary representation
of bytes are right-aligned (like big-endian numbers). To avoid thinking
of BitFlags as either big- or little-endian it may more appropriate to
speak of things in terms leading and trailing 0s.