Refine the test clean-up steps for `test_acquire_resource_on_drive()`.
On Windows with Python 3.7, this test case was triggering the following
exception:

    PermissionError: [WinError 32] The process cannot access the file
    because it is being used by another process

This exception was triggered because the `weakref.finalize()` function
(which removes the temporary database file) was running *before* the
test case's `addCleanup()` method. And since `addCleanup()` had not yet
closed the database connection, the file was still in use when the
finalizer tried to delete it.