Change some `Literal[..., None]` type hints to `Optional[...]`.
Important: This is only done for cases where it's semantically
correct to do so--not for all occurances where `None` is an
accepted literal value. It would be semantically incorrect to
characterize a `None` value using the `Optional` type hint when
there is no default value or when the default value is something
other than `None` (despite what any linter might suggest).