Change `update_index()` behavior to require all label columns.
Being able to load updates from a file that doesn't include all of a
node's label columns is a minor convenience. But mistakenly loading
updates from the wrong file (one that happens to use some of the same
column names) could be a major inconvenience if it wasn't immediately
noticed.