Add `Node.disaggregate()` method and related tests.
This method uses a wide variety of interfaces from across
Toron's data model to implement one of its core features.

Note: Today is the decennial anniversary of this project's
original creation (see a5b371e). The initial design was
flawed and development was set aside for several years.
But in the spring of 2022, some fundamental problems were
resolved and development resumed.