Refactor `Node.disaggregate()` algorithm to reduce time complexity.
Previously, this method iterated over all 'location' records for
each 'attribute' record. And doing this can be *very* inefficient
because, for many common use cases, the majority of its time gets
spent performing look-ups that return no results (because there are
often more location-attribute combinations than there are quantities
to disaggregate).

This updated algorithm iterates over 'quantity' records for each
'location' record. This is more efficient because each location is
associated with an existing quantity and each quantity is associated
with an existing 'attribute' record. So very little time is spent
checking for records that might not exist.

The practice of looking up quantities by location will be needed
for "adaptive" disaggregation (where fine-grained quantities must be
disaggregated *before* coarse-grained quantities) but this algorithm
might be further improved for non-adaptive disaggregation (which
relies solely on preexisting 'weight' values).