Change `Node.domain` setter into `Node.set_domain()` method.
This change was made to be consistent with the behavior of other
methods in the `Node` class. Other methods use the callable syntax
when altering persistent node data. Using callable syntax suggests
to the user that something more than simple label assignment is
happening behind the scenes--whereas the property setter obfuscates
this fact.