Update tests.yml to call an explicit list of tox environments.
Currently the GitHub Actions 'windows-latest' runner image still
has a version of Python 3.7 installed by default. Unfortunately,
it tries to use a version of `setuptools` that is incompatible
which causes an error. Since Python 3.7 has been end-of-life for
more than a year, this issue will not likely be fixed.

Calling 'tox' with an explicit list of environments allows
GitHub Actions to test Python versions that *it* supports and
allows developers to locally test versions that *they* support
while still using the same tox configuration.