Update disaggregation code to optimize for processing speed.
In testing, these changes sped-up processing by 5 - 15% for various
data sets. Changes that were made include:

* Eliminating one of the calls to `index_repo.find_by_label()`.
* Changing `data_service.disaggregate_value()` to accept a container
  of `index_id` values rather than looking them up in the local scope.

I tried a handful of ideas to optimize the disaggregation process
but some changes provided only marginal benefit (0 - 2% speed-up)
which was not worth the added complexity--at least at this time.

Some of the ideas that DID NOT provide a useful speed-up include:

* Eliminating repeated calls to `get_greatest_unique_specificity()`
  to determine `weight_group_id` matches for each quantity and
  replacing it with a pre-calculated cache.
* Eliminating all calls to `index_repo.find_by_label()` (which runs
  a SELECT query that preforms string matching in a WHERE clause) and
  replacing it with a pre-calculated lookup table of `_location_id`
  and `index_id` values that can be simply JOINed (even with added
  indexes, it didn't out-perform the single `find_by_label()` call).