Add `find_by_multiple()` method to abstract and concrete QuantityRepository classes.
While part of the "quantity" repository, this method must
also interact with the back-end "location" data. This steps
outside the normal scope of repository methods.

Using loosely coupled methods helps to keep the design
flexible, but it can limit the ability to implement data
access optimizations. In this project's critical loops,
we need to fetch many Quantity objects quickly, so
performance is important.