Remove location filtering from `find_by_multiple()` definition.
This is being removed for now because this method is used during
the disaggregation process. And quantities that are reported at a
more coarse-grained level cannot be accurately filtered by finer-
grained locations until *after* disaggregation has been performed.

For example, assume we have a node with state and county columns.
And we have quantities reported at the state level which represent
the population. We can't accurately filter this quantity to a
particular county until after the aggregate state-level value has
been disaggregated to the individual counties.