Add "type: ignore" to `NodeReader` to preserve EAFP style.
We want to stick with EAFP style because iteration should be as
fast as possible and we don't want to have to check the value of
`_data` on every iteration. Mypy does not cater to this style and
gives a spurious error.

See https://docs.python.org/3/glossary.html#term-EAFP