Add `translate3()` function which implements an alternative approach.
I hoped that we could achieve faster node-to-node translation by
pushing more of the work into the SQLite engine. But this did not
seem to help in real-world testing.

This function stores all data in SQLite databases and then uses
the ATTACH command to connect the `TopoNode` database (the DAL1
back-end), the source `NodeReader`, and the destination `NodeReader`
into a single database connection. The crosswalk's proportions are
then computed directly in a SQL query (rather than in a for-loop at
the application layer).

Unfortunately, this in-database approach was not materially faster.
When testing with real-world data, a translation that took 31 seconds
when handled at the application layer instead took 30 seconds when
handled at the database layer. Additionally, adding more database
indexes to the `TopoNode` and `NodeReader` databases did not improve
performance.

The one change that *did* improve performance for on-drive databases
was setting 'PRAGMA synchronous = OFF' (which is acceptable in the
case of these temporary NodeReader databases).

This function will be removed but it is included here to record its
implementation and its performance.