Change `pivot_reader_to_pandas()` to use empty string for missing values.
This change helps Toron's pivot to better mimic the behavior of
Pandas' `df.pivot_table()`. The behavior of `df.pivot_table()`
is to omit any columns with a NULL component and a common way to
assure that all columns are included is to replace NULL values
with empty strings (`df.loc[:, pivot_cols].fillna('')`).

A native Toron pivot was made because the Pandas `df.pivot_table()`
method is not memory efficient.