Optimize unmatched left-side handling in `_get_mapping_elements()`.
To yield unmatched left-side elements, each index in the left-side
node is checked individually at the application level. This is not
ideal, but it's the consequence of certain trade-offs made by the
current TopoNode design (loose coupling, decentralized graph, etc.).

However, it is possible to quickly determine if the left-side
is complete and up-to-date as a whole. This change compares the
left-node's hash against the crosswalk's hash--if they are the
same, then we know that there's no need to search for missing
left-side elements. This means that, the generator will finish
more quickly when there are no unmatched left-side elements.