Change `_get_ambiguous_fields()` to use `None` for unambiguous records.
Previously, empty string was returned for an unambiguous record.

This change (along with 866d2b4, 8f54f29, and d3001b3) makes the
output of `get_mapping()` more consistent. Processes that operate
on the output of `get_mapping()` can now use `is None` to check
for missing values in all cases.