Updated code comments for `create_sql_function()`.
The old comment mixed-up Python and SQLite versions.

To be clear, the `if` condition checks the *Python* version.
The optional `deterministic` argument is new in Python 3.8.0
(2019-10-14). However, the underlying database has supported
this functionality since SQLite 3.8.3 (2014-02-03). And Toron
currently supports SQLite 3.21.0 (2017-10-24) and newer--which
is the oldest SQLite packaged with Python 3.7.0 (2018-06-27).