Update `main()` function exception handling type.
The 'toron' command line interface should not intentionally emit
arbitrary, internal exceptions. To raise an error that should be
presented to the user, the `TorronError` type should be used.

The application layer (in 'main.py' for the CLI) should handle
`ToronError` instances and present error messages to the user with
appropriate formatting (ANSI colors, etc.).

When a non-ToronError exception leaks into the command line, it
should not be wrapped and presented to the user as a normal part
of operation. It should fail with all of the normal tracebacks
necessary to debug the issue.